% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsample-package.R
\docType{package}
\name{gsample-package}
\alias{gsample}
\alias{gsample-package}
\title{gsample: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vgherard/gsample}
  \item Report bugs at \url{https://github.com/vgherard/gsample/issues}
}

}
\keyword{internal}
