% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsample.R
\name{gsample}
\alias{gsample}
\alias{gsample.int}
\title{Efficient weighted sampling without replacement}
\usage{
gsample.int(n, size = n, replace = FALSE, prob = NULL)

gsample(x, size, replace = FALSE, prob)
}
\arguments{
\item{n}{length one integer. The total number of categories to choose from;
See ‘Details’.}

\item{size}{length one integer. Size of sample.}

\item{replace}{TRUE or FALSE. Sample with replacement? Defaults to
\code{FALSE}.}

\item{prob}{either \code{NULL}, or a numeric vector of length \code{n},
containing probability weights for sampling the various classes.
If \code{NULL} (default), sampling is performed assuming uniform
probabilities.}

\item{x}{a vector of one or more elements from which to choose.}
}
\value{
an integer vector of class indexes for \code{gsample.int}, a vector
of the same type of \code{x} for \code{gsample}.
}
\description{
Weighted sampling without replacement using the Gumbel-Max trick
(c.f. \url{https://arxiv.org/pdf/1903.06059.pdf}).
}
\examples{
set.seed(840)
gsample(letters, 5, prob = runif(length(letters)))
gsample.int(10, 3, prob = 10:1)
}
\author{
Valerio Gherardi
}
