% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsample-package.R
\docType{package}
\name{gsample-package}
\alias{gsample}
\alias{gsample-package}
\title{gsample: Efficient Weighted Sampling Without Replacement}
\description{
Sample without replacement using the Gumbel-Max trick (c.f. \url{https://arxiv.org/pdf/1903.06059.pdf}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vgherard/gsample}
  \item Report bugs at \url{https://github.com/vgherard/gsample/issues}
}

}
\keyword{internal}
