% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsample.R
\name{gsample}
\alias{gsample}
\alias{gsample.int}
\title{Efficient sampling without replacement}
\usage{
gsample.int(n, size = n, replace = FALSE, weights = NULL, log_weights = FALSE)

gsample(x, size, replace = FALSE, weights = NULL, log_weights = FALSE)
}
\arguments{
\item{n}{length one integer. The total number of categories to choose from;
See ‘Details’.}

\item{size}{length one integer. Size of sample.}

\item{replace}{TRUE or FALSE. Sample with replacement? Defaults to
\code{FALSE}.}

\item{weights}{either \code{NULL}, or a numeric vector of length \code{n},
containing the (unnormalized) probabilities, or the (unnormalized)
log-probabilities for sampling categories
(depending on the value of \code{log_weights} argument, see below).
If \code{NULL} (default), sampling is performed assuming uniform
probabilities.}

\item{log_weights}{TRUE or FALSE. Should the values given in \code{weights}
be interpreted as (unnormalized) log-probabilities? Defaults to \code{FALSE}.}
}
\value{
an integer vector if \code{map} is \code{NULL}, otherwise a vector
of the same type as \code{map}.
}
\description{
Sample without replacement using the Gumbel-Max trick
(c.f. \url{https://arxiv.org/pdf/1903.06059.pdf}).
}
\details{

}
\examples{
set.seed(840)
size <- 5
weights <- rnorm(length(letters), sd = 10)
gsample(size = 5, weights, map = letters)
}
\author{
Valerio Gherardi
}
