% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsample.R
\name{gsample}
\alias{gsample}
\title{Sample a discrete random variable given its log-probabilities}
\usage{
gsample(size, weights, replace = FALSE, log_weights = FALSE, map = NULL)
}
\arguments{
\item{size}{length one integer. Size of sample.}

\item{weights}{a numeric vector, containing either (unnormalized)
probabilities or (unnormalized) log-probabilities of categories
(c.f. description of \code{log_weights} argument).}

\item{replace}{TRUE or FALSE. Sample with replacement? Defaults to
\code{FALSE}.}

\item{log_weights}{TRUE or FALSE. Are the \code{weights} given as
(unnormalized) log-probabilities? Defaults to \code{FALSE}.}

\item{map}{either \code{NULL} or a vector of length \code{length(weights)}.
If \code{NULL}, the returned value is an integer vector containing the
indexes of the sampled classes, otherwise the elements of \code{map}
corresponding to such indexes.}
}
\value{
an integer vector if \code{map} is \code{NULL}, otherwise a vector
of the same type as \code{map}.
}
\description{
Sample from a discrete probability distribution with Pr(i) ∝ exp(phi\link{i})
using the Gumbel-Max trick (c.f. \url{https://lips.cs.princeton.edu/the-gumbel-max-trick-for-discrete-distributions/}).
}
\author{
Valerio Gherardi
}
